Fix error handling and paging, improve handling of tasks stops and deletions, load pending sites on server start

+++ src/plugins/task/task.service.ts
@@ -9,6 +9,7 @@ import { Task } from '@model/domain/task';
 import { TaskConfigDto, LocalTaskConfig } from '@model/dto/task-config.dto';
 import { TaskRunHandle } from '@modules/contract/model/task-run-handle';
 import { TaskCallbackFn } from '@plugins/task/model/task-callback-fn';
+import { ExecuteTaskFn } from '@modules/contract/model/task-queue.contract';
 
 export class TaskService {
   private readonly logger = getLogger('task-service');
@@ -20,21 +21,36 @@ export class TaskService {
   ) {}
 
   async registerFromDatabase(): Promise<void> {
-    const tasks = await this.dao.findAll();
-    const activeTasks = tasks.filter((task) => task.active);
-    for (const task of activeTasks) {
-      const { id, config } = task;
+    const tasks = await this.dao.findAllByActive(true);
+    for (const task of tasks) {
+      const { id: taskId, config } = task;
+      await this.enqueuePendingRuns(task);
+
       const taskConfig: TaskConfigDto = JSON.parse(config);
-      await this.register(id, taskConfig);
+      await this.register(taskId, taskConfig);
     }
+    this.logger.info(`registerFromDatabase: registered ${tasks.length} tasks`);
+  }
+
+  private async enqueuePendingRuns(task: Task): Promise<void> {
+    const { id: taskId } = task;
+    const pendingRuns = await this.runService.getAllByStatus(taskId, 'pending');
+
+    for (const taskRun of pendingRuns) {
+      const { id: runId, rank: runRank } = taskRun;
+      const runHandle: TaskRunHandle = { taskId, runId, runRank };
+      const executorFn = this.buildExecutorFn(runHandle);
+      await this.executorService.enqueueExecutable(taskId, executorFn);
+    }
+
     this.logger.info(
-      `registerFromDatabase: registered ${activeTasks.length}/${tasks.length} tasks`
+      `enqueuePendingRuns[taskId=${taskId}]: ${pendingRuns.length} runs`
     );
   }
 
   async register(taskId: number, taskConfig: TaskConfigDto): Promise<void> {
     const callbackFn = this.buildTaskCallbackFn();
-    await this.executorService.registerExecutable(
+    await this.executorService.scheduleExecutable(
       taskId,
       taskConfig,
       callbackFn
@@ -44,47 +60,54 @@ export class TaskService {
 
   private buildTaskCallbackFn(): TaskCallbackFn {
     return async (taskId) => {
-      const task = await this.dao.findById(taskId); // to handle updates/deletions
-      if (task === undefined) {
-        this.logger.warn(`executorFn[taskId=${taskId}]: SKIP - doesn't exist`);
+      const [isActive, task] = await this.isTaskActive(taskId);
+      if (!isActive) {
+        this.logger.warn(`callbackFn[taskId=${taskId}]: skip - is not allowed`);
         return null;
       }
-      const config: TaskConfigDto = JSON.parse(task.config);
-      const [runHandle, runIndex] = await this.runService.tryToRegister(
-        taskId,
-        config.times
-      );
-      if (runHandle === null) {
+      const config: TaskConfigDto = JSON.parse((task as Task).config);
+      const taskRun = await this.runService.tryToRegister(taskId, config.times);
+      if (taskRun === null) {
         await this.executorService.unscheduleExecutable(taskId);
         return null;
       }
-      return () => this.runTask(runHandle, runIndex, config);
+      const runHandle: TaskRunHandle = {
+        taskId,
+        runId: taskRun.id,
+        runRank: taskRun.rank,
+      };
+      return this.buildExecutorFn(runHandle);
     };
   }
 
-  private async runTask(
-    runHandle: TaskRunHandle,
-    runIndex: number,
-    taskConfig: TaskConfigDto
-  ): Promise<void> {
+  private buildExecutorFn(runHandle: TaskRunHandle): ExecuteTaskFn {
+    return () => this.runTask(runHandle);
+  }
+
+  private async runTask(runHandle: TaskRunHandle): Promise<void> {
     try {
-      await this.runTaskImpl(runHandle, runIndex, taskConfig);
+      await this.runTaskImpl(runHandle);
     } catch (err) {
-      const { taskId } = runHandle;
+      const { taskId, runId } = runHandle;
       const reason = errorToReason(err);
-      this.logger.warn(`runTask[taskId=${taskId}]: failed - ${reason}`);
+      this.logger.warn(
+        `runTask[taskId=${taskId}, runId=${runId}]: failed - ${reason}`
+      );
       await this.stopTask(taskId);
     }
   }
 
   // task is like a clean function, so there is no need to re-build it
-  private async runTaskImpl(
-    runHandle: TaskRunHandle,
-    runIndex: number,
-    taskConfig: TaskConfigDto
-  ): Promise<void> {
+  private async runTaskImpl(runHandle: TaskRunHandle): Promise<void> {
     const { taskId } = runHandle;
-    const { times, config } = taskConfig;
+    const [isRunAllowed, task] = await this.isTaskActive(taskId);
+    if (!isRunAllowed) {
+      this.logger.warn(`runTask[taskId=${taskId}]: skip - is not allowed`);
+      return;
+    }
+
+    const taskConfig: TaskConfigDto = JSON.parse((task as Task).config);
+    const { config, times } = taskConfig;
 
     const envHandle = await this.executorService.runExecutable(taskId, config);
     await this.runService.start(runHandle, envHandle);
@@ -92,13 +115,27 @@ export class TaskService {
     await this.runService.finish(runHandle, exitCode === 0);
 
     await envHandle.delete({ isForce: false });
-    this.logger.info(`runTask[taskId]: completed - exitCode=${exitCode}`);
+    this.logger.info(`runTask[taskId=${taskId}]: exitCode=${exitCode}`);
 
-    if (runIndex === times) {
-      await this.stopTask(taskId);
+    if (times !== undefined && runHandle.runRank >= times) {
+      await this.executorService.stopExecutable(taskId);
       await this.executorService.deleteExecutableEnv(taskId);
+      await this.dao.setActive(taskId, false);
       this.logger.info(`runTask[taskId=${taskId}]: stop - no runs left`);
     }
+  }
+
+  private async isTaskActive(taskId: number): Promise<[boolean, Task | null]> {
+    const task = await this.dao.findById(taskId);
+    if (task === undefined) {
+      this.logger.warn(`isTaskActive[taskId=${taskId}]: does not exist`);
+      return [false, null];
+    }
+    if (!task.active) {
+      this.logger.warn(`isTaskActive[taskId=${taskId}]: is NOT active`);
+      return [false, task];
+    }
+    return [true, task];
   }
 
   async createTask(
@@ -106,15 +143,15 @@ export class TaskService {
     userId: number,
     taskConfig: TaskConfigDto
   ): Promise<number> {
-    const { name } = taskConfig;
-    if ((await this.dao.findByName(name)) !== undefined) {
-      throw new EntityConflict(`Task with name=${name} already exists`);
+    const uniqueName = buildTaskUniqueName(groupId, taskConfig.name);
+    if ((await this.dao.findByName(uniqueName)) !== undefined) {
+      throw new EntityConflict(`Task with name=${uniqueName} already exists`);
     }
 
     const taskId = await this.dao.create({
       userId,
       groupId,
-      name: buildTaskUniqueName(groupId, name),
+      name: uniqueName,
       active: true,
       config: JSON.stringify(taskConfig),
     });
@@ -125,7 +162,11 @@ export class TaskService {
   }
 
   async deleteTask(taskId: number): Promise<void> {
-    await this.dao.findByIdThrowable(taskId);
+    const task = await this.dao.findByIdThrowable(taskId);
+    if (task.active) {
+      this.logger.warn(`deleteTask[taskId=${taskId}]: task is active`);
+      throw new BadRequest(`Cannot delete active task: taskId=${taskId}`);
+    }
     await this.executorService.deleteExecutable(taskId);
     await this.runService.flushAll(taskId);
     await this.dao.delete(taskId);
@@ -136,10 +177,11 @@ export class TaskService {
     const task = await this.dao.findByIdThrowable(taskId);
     if (!task.active) {
       this.logger.warn(`stopTask[taskId=${taskId}]: task is NOT active`);
-      throw new BadRequest(`Task is not active: taskId=${taskId}`);
+      throw new BadRequest(`Cannot stop NOT active task: taskId=${taskId}`);
     }
-    await this.executorService.stopExecutable(taskId);
     await this.dao.setActive(taskId, false);
+    await this.executorService.stopExecutable(taskId);
+    await this.runService.stopAll(taskId);
     this.logger.info(`stopTask[taskId=${taskId}]: set active=false`);
   }
 
@@ -147,7 +189,7 @@ export class TaskService {
     const task = await this.dao.findByIdThrowable(taskId);
     if (task.active) {
       this.logger.warn(`startTask[taskId=${taskId}]: task is active`);
-      throw new BadRequest(`Task is already active: taskId=${taskId}`);
+      throw new BadRequest(`Cannot start active task: taskId=${taskId}`);
     }
     const taskConfig = JSON.parse(task.config);
     await this.register(taskId, taskConfig);
